+++ src/lspClient/middlewares/codeActionMiddleware.ts
@@ -2,6 +2,15 @@ import * as vscode from 'vscode';
 import * as vslc from 'vscode-languageclient';
 import { CommandIds } from '../../commandIds';
 
+/**
+ * Sorbet specifically requires this command Id which leads to compatibility
+ * complexities when multiple extensions want to host Sorbet as they need to
+ * offer the command. To workaround this, this middleware remaps the command
+ * to extension-owned command.
+ * https://sorbet.org/docs/lsp#sorbetsavepackagefiles-command
+ */
+const SORBET_SAVE_PACKAGE_FILES = 'sorbet.savePackageFiles';
+
 export const CodeActionMiddleware: vslc.CodeActionMiddleware = {
   provideCodeActions: async (
     document: vscode.TextDocument,
@@ -11,6 +20,11 @@ export const CodeActionMiddleware: vslc.CodeActionMiddleware = {
     next: vslc.ProvideCodeActionsSignature,
   ): Promise<(vscode.Command | vscode.CodeAction)[] | undefined | null> => {
     const actions = (await next(document, range, context, token)) ?? [];
+    actions.forEach(({ command }) => {
+      if (command && typeof command === 'object' && command.command === SORBET_SAVE_PACKAGE_FILES) {
+        command.command = CommandIds.SavePackageFiles;
+      }
+    });
 
     const diagnostic = context.diagnostics.find(
       (d) =>